+++ src/main/java/com/uitopic/restock/platform/tracking/domain/model/valueobjects/StockRecord.java
@@ -30,11 +30,11 @@ public record StockRecord(
      * @param threshold the maximum allowed difference between the two stock records for them to be considered within the threshold
      * @return true if the absolute difference between the two stock records is less than or equal to the threshold, false otherwise
      */
-    public Boolean isInThreshold(StockRecord other, Integer threshold) {
+    public Boolean isInThreshold(StockRecord other, Double threshold) {
         if (other == null) {
             throw new StockComparisonIncompletedException("Stock to compare cannot be null");
         }
-        if (threshold == null || threshold < 0) {
+        if (threshold == null || threshold < 0.0) {
             throw new StockComparisonIncompletedException("Threshold cannot be null or negative");
         }
         return Math.abs(this.stock - other.stock) <= threshold;
